Minimally viable functionality (#1)
Minimally viable functionality has been built and is ready to be merged into master.

This PR includes a working /stock endpoint from where web clients can read the current product stock by issuing a GET request to this endpoint.

The web server is configured via environment variables and provides a Dockerfile for building and deploying the application instance in small Alpine Linux based Docker containers.

There is not authentication built into the server at this time but authentication middleware has been written, just not integrated or tested yet. This will be done in the next version.